Convert cvm::to_str() from a template to a fixed set of overloaded functions
By instantiating the stringification template only once per overloaded
function, this reduces the size of nearly all object files, except
colvarbias_abf.o, for some reason.

+++ src/colvarbias.cpp
@@ -338,7 +338,8 @@ std::istream & colvarbias::read_state(std::istream &is)
        (set_state_params(conf) != COLVARS_OK) ) {
     cvm::error("Error: in reading state configuration for \""+bias_type+"\" bias \""+
                this->name+"\" at position "+
-               cvm::to_str(is.tellg())+" in stream.\n", INPUT_ERROR);
+               cvm::to_str(static_cast<size_t>(is.tellg()))+
+               " in stream.\n", INPUT_ERROR);
     is.clear();
     is.seekg(start_pos, std::ios::beg);
     is.setstate(std::ios::failbit);
@@ -348,7 +349,8 @@ std::istream & colvarbias::read_state(std::istream &is)
   if (!read_state_data(is)) {
     cvm::error("Error: in reading state data for \""+bias_type+"\" bias \""+
                this->name+"\" at position "+
-               cvm::to_str(is.tellg())+" in stream.\n", INPUT_ERROR);
+               cvm::to_str(static_cast<size_t>(is.tellg()))+
+               " in stream.\n", INPUT_ERROR);
     is.clear();
     is.seekg(start_pos, std::ios::beg);
     is.setstate(std::ios::failbit);
@@ -358,7 +360,8 @@ std::istream & colvarbias::read_state(std::istream &is)
   if (brace != "}") {
     cvm::error("Error: corrupt restart information for \""+bias_type+"\" bias \""+
                this->name+"\": no matching brace at position "+
-               cvm::to_str(is.tellg())+" in stream.\n");
+               cvm::to_str(static_cast<size_t>(is.tellg()))+
+               " in stream.\n");
     is.setstate(std::ios::failbit);
   }
 
@@ -374,7 +377,8 @@ std::istream & colvarbias::read_state_data_key(std::istream &is, char const *key
        !(key_in == to_lower_cppstr(std::string(key))) ) {
     cvm::error("Error: in reading restart configuration for "+
                bias_type+" bias \""+this->name+"\" at position "+
-               cvm::to_str(is.tellg())+" in stream.\n", INPUT_ERROR);
+               cvm::to_str(static_cast<size_t>(is.tellg()))+
+               " in stream.\n", INPUT_ERROR);
     is.clear();
     is.seekg(start_pos, std::ios::beg);
     is.setstate(std::ios::failbit);
